+++ index.html
@@ -1,101 +1,114 @@
 <!DOCTYPE html>
 <html>
+  <head>
+    <title>Проект "Комменты"</title>
+    <meta charset="utf-8" />
+    <link rel="icon" href="./favicon.ico" type="image/x-icon" />
+    <link rel="stylesheet" href="styles.css" />
+  </head>
 
-<head>
-  <title>Проект "Комменты"</title>
-  <meta charset="utf-8" />
-  <link rel="icon" href="./favicon.ico" type="image/x-icon">
-  <link rel="stylesheet" href="styles.css" />
-</head>
-
-<body>
-  <div class="container">
-    <div class="list-loader-get">
-      <div class="windows8">
-        <div class="wBall" id="wBall_1">
-          <div class="wInnerBall"></div>
+  <body>
+    <!-- <div class="container">
+      <div class="list-loader-get">
+        <div class="windows8">
+          <div class="wBall" id="wBall_1">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_2">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_3">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_4">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_5">
+            <div class="wInnerBall"></div>
+          </div>
         </div>
-        <div class="wBall" id="wBall_2">
-          <div class="wInnerBall"></div>
+      </div>
+      <ul class="comments"></ul>
+      <div class="add-form">
+        <input
+          type="text"
+          class="add-form-name"
+          placeholder="Введите ваше имя"
+        />
+        <textarea
+          type="textarea"
+          class="add-form-text"
+          placeholder="Введите ваш коментарий"
+          rows="4"
+        ></textarea>
+        <div class="add-form-row">
+          <button class="add-form-button">Написать</button>
+          <button class="add-form-button delete-form-button">
+            Удалить последний комментарий
+          </button>
         </div>
-        <div class="wBall" id="wBall_3">
-          <div class="wInnerBall"></div>
+      </div>
+      <div class="list-loader-post hidden">
+        <div class="windows8">
+          <div class="wBall" id="wBall_1">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_2">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_3">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_4">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_5">
+            <div class="wInnerBall"></div>
+          </div>
         </div>
-        <div class="wBall" id="wBall_4">
-          <div class="wInnerBall"></div>
+      </div>
+    </div> -->
+    <div class="container">
+      <div class="list-loader-get hidden">
+        <div class="windows8">
+          <div class="wBall" id="wBall_1">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_2">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_3">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_4">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_5">
+            <div class="wInnerBall"></div>
+          </div>
         </div>
-        <div class="wBall" id="wBall_5">
-          <div class="wInnerBall"></div>
+      </div>
+      <div class="app"></div>
+      <div class="list-loader-post hidden">
+        <div class="windows8">
+          <div class="wBall" id="wBall_1">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_2">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_3">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_4">
+            <div class="wInnerBall"></div>
+          </div>
+          <div class="wBall" id="wBall_5">
+            <div class="wInnerBall"></div>
+          </div>
         </div>
       </div>
     </div>
-    <ul class="comments">
-      <!-- Список коментариев рендерится из JS -->
-      <!-- <li class="comment">
-        <div class="comment-header">
-          <div>Глеб Фокин</div>
-          <div>12.02.22 12:18</div>
-        </div>
-        <div class="comment-body">
-          <div class="comment-text">
-            Это будет первый комментарий на этой странице
-          </div>
-        </div>
-        <div class="comment-footer">
-          <div class="likes">
-            <span class="likes-counter">3</span>
-            <button class="like-button"></button>
-          </div>
-        </div>
-      </li>
-      <li class="comment">
-        <div class="comment-header">
-          <div>Варвара Н.</div>
-          <div>13.02.22 19:22</div>
-        </div>
-        <div class="comment-body">
-          <div class="comment-text">
-            Мне нравится как оформлена эта страница! ❤
-          </div>
-        </div>
-        <div class="comment-footer">
-          <div class="likes">
-            <span class="likes-counter">75</span>
-            <button class="like-button -active-like"></button>
-          </div>
-        </div>
-      </li> -->
-    </ul>
-    <div class="add-form">
-      <input type="text" class="add-form-name" placeholder="Введите ваше имя" />
-      <textarea type="textarea" class="add-form-text" placeholder="Введите ваш коментарий" rows="4"></textarea>
-      <div class="add-form-row">
-        <button class="add-form-button">Написать</button>
-        <button class="add-form-button delete-form-button">Удалить последний комментарий</button>
-      </div>
-    </div>
-    <div class="list-loader-post hidden">
-      <div class="windows8">
-        <div class="wBall" id="wBall_1">
-          <div class="wInnerBall"></div>
-        </div>
-        <div class="wBall" id="wBall_2">
-          <div class="wInnerBall"></div>
-        </div>
-        <div class="wBall" id="wBall_3">
-          <div class="wInnerBall"></div>
-        </div>
-        <div class="wBall" id="wBall_4">
-          <div class="wInnerBall"></div>
-        </div>
-        <div class="wBall" id="wBall_5">
-          <div class="wInnerBall"></div>
-        </div>
-      </div>
-    </div>
-  </div>
-  <div class="app"></div>
-  <script type="module" src="./main.js"></script>
-</body>
-
+    <script type="module" src="./main.js"></script>
+  </body>
 </html>
