+++ main.js
@@ -9,12 +9,6 @@ import {
 import { renderComments } from "./renderComments.js";
 import { renderLogin } from "./renderLogin.js";
 
-// Объявляем глобальные константы для всего проекта
-const inputName = document.querySelector(".add-form-name");
-const inputText = document.querySelector(".add-form-text");
-const addButtonElement = document.querySelector(".add-form-button");
-const addFormElement = document.querySelector(".add-form");
-
 // Храним информацию о пользователях в массиве
 let persons = [];
 
@@ -43,8 +37,8 @@ const getCommentsInfo = () => {
       });
 
       persons = appComments;
-      renderComments({ persons, inputText });
-      initDeleteLastComentListener({ persons, renderComments, inputText });
+      renderComments({ persons, postCommentInfo });
+      initDeleteLastComentListener({ persons, renderComments, postCommentInfo });
     })
     .then(() => {
       hideListLoaderGet();
@@ -59,11 +53,10 @@ const getCommentsInfo = () => {
     });
 };
 
-renderLogin({ getCommentsInfo });
 // getCommentsInfo();
 
 // Функция для добавления данных о пользователе в БД API
-const postCommentInfo = () => {
+const postCommentInfo = ({inputText, inputName}) => {
   return postComment({
     inputText: inputText.value.vulnerabilityPrevention(),
     inputName: inputName.value.vulnerabilityPrevention(),
@@ -82,80 +75,5 @@ const postCommentInfo = () => {
     });
 };
 
-// Кнопка "Написать" будет недоступна прям в момент посещения сайта, так как поля пустые
-// в дальнейшем кнопка разблокируется, если заполнить поля каким-либо текстом
-addButtonElement.disabled = true;
+renderLogin({ getCommentsInfo });
 
-// Тут кнопка разблокируется, если в какой либо форме будет текст
-inputName.addEventListener("input", () => {
-  if (inputName.value.length > 0) {
-    return (addButtonElement.disabled = false);
-  } else {
-    return (addButtonElement.disabled = true);
-  }
-});
-
-inputText.addEventListener("input", () => {
-  if (inputText.value.length > 0) {
-    return (addButtonElement.disabled = false);
-  } else {
-    return (addButtonElement.disabled = true);
-  }
-});
-
-// Валидация форм. Если какое либо поле будет пустым, placeholder окраситься в красный
-// Если валидация пройдена, коментарий добавиться в ленту по нажатию мышкой на кнопку
-// После того, как коментарий опубликуется, произойдет очистка полей
-addButtonElement.addEventListener("click", () => {
-  inputName.classList.remove("error");
-  inputText.classList.remove("error");
-
-  if (inputName.value.isEmpty() || inputText.value.isEmpty()) {
-    inputText.value = "";
-    inputName.value = "";
-    inputName.classList.add("error");
-    inputText.classList.add("error");
-    return;
-  }
-
-  addFormElement.style.display = "none";
-  showListLoaderPost();
-
-  postCommentInfo().then(() => {
-    addFormElement.style.display = "flex";
-    hideListLoaderPost();
-  });
-});
-
-// То же самое, что и выше, только для добавления коментария по нажатию "Enter"
-inputText.addEventListener("keyup", (event) => {
-  if (event.key === "Enter") {
-    // Проверяем, была ли нажата клавиша Enter
-
-    inputName.classList.remove("error");
-    inputText.classList.remove("error");
-
-    if (
-      inputName.value.isEmpty() ||
-      inputText.value.isEmpty() ||
-      inputText.value === "\n"
-    ) {
-      inputText.value = "";
-      inputName.value = "";
-      inputName.classList.add("error");
-      inputText.classList.add("error");
-      return;
-    }
-
-    addFormElement.style.display = "none";
-    showListLoaderPost();
-
-    postCommentInfo().then(() => {
-      addFormElement.style.display = "flex";
-      hideListLoaderPost();
-    });
-
-    // Вызываем рендер функцию для отрисовки нового коментария
-    renderComments({ persons, inputText });
-  }
-});
